# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateUser:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.EnableUser:
    properties:
      enabled:
        type: boolean
    type: object
  models.Forgot_User_Password_Payload:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.Permission:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.SendMail:
    properties:
      email:
        type: string
    type: object
  models.Set_User_Password_Payload:
    properties:
      password:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      provider:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      provider:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    name: API Support
  description: This is an RBAC based full fledge API serice for managing users, roles
    and permissions in the system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Truvest Identity Management Service APIs
  version: "1.0"
paths:
...
  /user/me:
    get:
      consumes:
      - application/json
      description: Get the logged in user details by Authorization header being passed.
        This is useful by other microservices to check for authorization based upon
        the roles and permissions listed.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Get the Logged in User details
      tags:
      - User
  /users:
    get:
      consumes:
      - application/json
      description: Get details of all users in the system. In order to access this
        API, someone must have "SYSTEM_ADMIN" and "USERS_VIEW" Permission tagged to
        its role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - ApiKeyAuth: []
      summary: Get all users in the system
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a user by ID in the system. In order to access this API,
        someone must have "USERS_CREATE" Permission tagged to its role. If Password
        is not passed while creating User, then it creates User with auto-generated
        password and sends it through Email.
      parameters:
      - description: Create User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a user in the system
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID in the system. In order to access this API,
        someone must have "USERS_CREATE" Permission tagged to its role.
      parameters:
      - description: ID of the user to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a user in the system
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get details of a user by ID in the system. In order to access this
        API, someone must have "USERS_VIEW" Permission tagged to its role.
      parameters:
      - description: ID of the user to get
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a user by ID in the system
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update a user by ID in the system. In order to access this API,
        someone must have "USERS_CREATE" Permission tagged to its role.
      parameters:
      - description: ID of the user to be updated
        in: path
        name: id
        required: true
        type: integer
      - description: Update User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a user in the system
      tags:
      - User
  /users/{id}/enableUser:
    put:
      consumes:
      - application/json
      description: Enable/Disable a user by ID in the system. In order to access this
        API, someone must have "USERS_CREATE" Permission tagged to its role. Set Enabled
        field in payload to true to enable a user/Set Enabled field in payload to
        false to disable a user
      parameters:
      - description: ID of the user to be updated
        in: path
        name: id
        required: true
        type: integer
      - description: Enable User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.EnableUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Enable/Disable a user in the system
      tags:
      - User
  /users/{id}/setPassword:
    post:
      consumes:
      - application/json
      description: Update a user's password by ID in the system. The User himself
        can only change the password through this endpoint.
      parameters:
      - description: ID of the user to be updated
        in: path
        name: id
        required: true
        type: string
      - description: Update User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Set_User_Password_Payload'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Set the Password of the user.
      tags:
      - User
  /users/forgotPassword:
    post:
      consumes:
      - application/json
      description: Set the Password of the user if he has forgotten.
      parameters:
      - description: Forgot Password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Forgot_User_Password_Payload'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Set the Password of the user if he has forgotten
      tags:
      - User
  /users/sendMail:
    post:
      consumes:
      - application/json
      description: Send a mail with a link for the user to reset password. This can
        be used in combination with Forget Password API to make sure a legitimate
        user is trying to reset his password.
      parameters:
      - description: Send Email
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SendMail'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Send a mail with a link for the user to reset password
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"